+++ src/plugins/dashboard/public/application/components/dashboard_variables/variables_bar.tsx
@@ -19,7 +19,6 @@ import {
   EuiPanel,
   EuiLoadingSpinner,
   EuiIconTip,
-  EuiFieldText,
 } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
 import { VariableService } from '../../../variables/variable_service';
@@ -28,6 +27,7 @@ import {
   buildVariableOptionDisplayTextMap,
   getVariableOptionDisplayText,
 } from '../../../variables/variable_option_display_utils';
+import { TextValueEditor } from './text_value_editor';
 import './variable_selector.scss';
 
 export interface VariablesBarProps {
@@ -376,63 +376,6 @@ const ValueSelector: React.FC<ValueSelectorProps> = ({ variable, onValuesChange 
         )}
       </EuiSelectable>
     </EuiPopover>
-  );
-};
-
-/**
- * Free-text input for Text-type variables.
- * Commits the value on blur or Enter so dependent query variables
- * only re-run once the user has finished typing.
- */
-const TextValueEditor: React.FC<ValueSelectorProps> = ({ variable, onValuesChange }) => {
-  const committedValue = variable.current?.[0] ?? '';
-  const [draft, setDraft] = useState(committedValue);
-
-  useEffect(() => {
-    setDraft(committedValue);
-  }, [committedValue]);
-
-  const commit = useCallback(() => {
-    if (draft !== committedValue) {
-      onValuesChange(variable.id, draft ? [draft] : []);
-    }
-  }, [draft, committedValue, variable.id, onValuesChange]);
-
-  const displayLabel = variable.label || variable.name;
-  // Base width on the committed value, not the in-progress draft — sizing off
-  // draft would make the box jitter on every keystroke.
-  const calculatedWidth = Math.max(
-    120,
-    Math.min(Math.max(committedValue.length, displayLabel.length) * 8 + 40, 400)
-  );
-
-  return (
-    <EuiToolTip content={variable.description} position="bottom">
-      <div
-        className="variableSelectorContainer"
-        data-label={displayLabel}
-        data-test-subj={`variable-${variable.name}`}
-        style={{ width: `${calculatedWidth}px` }}
-      >
-        <EuiFieldText
-          value={draft}
-          onChange={(e) => setDraft(e.target.value)}
-          onBlur={commit}
-          onKeyDown={(e) => {
-            if (e.key === 'Enter') {
-              commit();
-            }
-          }}
-          placeholder={i18n.translate('dashboard.variables.textPlaceholder', {
-            defaultMessage: 'Enter value...',
-          })}
-          data-test-subj="variable-text-input"
-          className="variableTextInput"
-          compressed
-          controlOnly
-        />
-      </div>
-    </EuiToolTip>
   );
 };
